minimal refactoring for containers and DataSet object
