+++ peepingtom/core/__init__.py
@@ -1,7 +1,21 @@
 from .alchemists import Alchemist, PointToLineAlchemist, PointToParticleAlchemist
-from .datablocks import DataBlock, SimpleBlock, MultiBlock, DataCrate, PointBlock, \
-    LineBlock, OrientationBlock, ImageBlock, LineBlock, SphereBlock, PropertyBlock, \
-    ParticleBlock, MeshBlock, DipoleBlock, TransformBlock
+from .datablocks import (
+    DataBlock,
+    SimpleBlock,
+    MultiBlock,
+    PointBlock,
+    LineBlock,
+    OrientationBlock,
+    ImageBlock,
+    LineBlock,
+    SphereBlock,
+    PropertyBlock,
+    ParticleBlock,
+    MeshBlock,
+    DipoleBlock,
+    TransformBlock,
+)
+from .containers import DataCrate, DataSet
 from .models import Model
 
 from .functions import merge, stack
